requirements.txt
python-dotenv
pandas

azure-core
azure-identity
azure-purview-catalog
azure-purview-administration